Add files via upload
A powerful fork bomb for use with an Arduino based rubber ducky USB. Perfectly functional with 'bad beetle' type expansion boards, programmed with Arduino IDE under board type Leonardo. This piece of code is free for use by anyone with a basic understanding of cyber law, and is not to be used under any circumstances on other people's computers without their consent. Or even with. This has the ability to damage a computers CPU. This was made to show the weaknesses in a windows operated machine without specific protection from this kind of attack, and is not meant to be used as a prank. Credit to It'sN8 for the CPU intensive side of the fork bomb, while I designed the rest. This fork bomb has a built in delay from startup, and then a random number generator to decide whether or not to run itself. It then has a small chance to use a different path to waste computer resources, impacting processing rather than RAM. The file will run anywhere from 1 minute to 20 minutes after the host starts up, a 1/6.25 chance to actually run, and then a 1/5.333 chance to run a more damaging form of the code. Reiteration: no pranks- pen-testing only. (for legal reasons)