remix of creditcard. programable via arduino ide.
this script disables win.def for C:/, adds a specific exception to startup, creates a scheduled startup task and creates System.vbs, which is an annoying script that cannot be stopped without using task manager. this was built for educational purposes, specifically for testing area of weakness is windows security. i take no responsibility for the use of this script outside of its intended area. use at your own risk.